+++ DoIt/database/CardDataBase.cpp
@@ -9,7 +9,7 @@ json CardDataBase::addCard(const json &info) const {
     return response;
 }
 
-json CardDataBase::removeCard(const int id) const {
+json CardDataBase::removeCard(const size_t id) const {
     json response = client->remove({{"id", id}});
     return response;
 }
@@ -19,22 +19,22 @@ json CardDataBase::updateCard(const json &info) const {
     return response;
 }
 
-json CardDataBase::getCardInfo(const int id) const {
+json CardDataBase::getCardInfo(const size_t id) const {
     json response = client->select({{"id", id}});
     return response;
 }
 
-json CardDataBase::getCardCheckLists(const int id) const {
+json CardDataBase::getCardCheckLists(const size_t id) const {
     json response = client->select({{"id", id}});
     return response;
 }
 
-json CardDataBase::getCardTags(const int id) const {
+json CardDataBase::getCardTags(const size_t id) const {
     json response = client->select({{"id", id}});
     return response;
 }
 
-json CardDataBase::getCardColumn(const int id) const {
+json CardDataBase::getCardColumn(const size_t id) const {
     json response = client->select({{"id", id}});
     return response;
 }
@@ -44,7 +44,7 @@ json CardDataBase::addTag(const json &info) const {
     return response;
 }
 
-json CardDataBase::removeTag(const int id) const {
+json CardDataBase::removeTag(const size_t id) const {
     json response = client->insert({{"id", id}});
     return response;
 }
@@ -54,7 +54,7 @@ json CardDataBase::addCheckList(const json &info) const {
     return response;
 }
 
-json CardDataBase::removeCheckList(const int id) const {
+json CardDataBase::removeCheckList(const size_t id) const {
     json response = client->insert({{"id", id}});
     return response;
 }
@@ -64,7 +64,7 @@ json CardDataBase::addCheckListItem(const json &info) const {
     return response;
 }
 
-json CardDataBase::removeCheckListItem(const int id) const {
+json CardDataBase::removeCheckListItem(const size_t id) const {
     json response = client->insert({{"id", id}});
     return response;
 }
