+++ DoIt/database/UserDataBase.h
@@ -10,11 +10,11 @@ class UserDataBase {
     UserDataBase() = default;
     UserDataBase(std::shared_ptr<DataBase> client);
     json addUser(const json &info) const;
-    json removeUser(const int id) const;
-    bool checkUserExists(const int id) const;
+    json removeUser(const size_t id) const;
+    bool checkUserExists(const size_t id) const;
     json updateUser(const json &info) const;
-    json getUserInfo(const int id) const;
-    json getUserBoards(const int id) const;
+    json getUserInfo(const size_t id) const;
+    json getUserBoards(const size_t id) const;
 
  private:
     std::shared_ptr<DataBase> client;
